Removed Food.java and replaced with MainItem.java, SideItem.java, and DrinkItem to reflect changes in design.
